Add more tests for optimiser
