+++ hypothesis-python/tests/conjecture/test_conjecture_engine.py
@@ -1031,32 +1031,6 @@ def test_prefix_cannot_exceed_buffer_size(monkeypatch):
         assert runner.valid_examples == buffer_size
 
 
-def test_optimises_multiple_targets():
-    with deterministic_PRNG():
-
-        def test(data):
-            n = data.draw_bits(8)
-            m = data.draw_bits(8)
-            if n + m > 256:
-                data.mark_invalid()
-            data.target_observations["m"] = m
-            data.target_observations["n"] = n
-            data.target_observations["m + n"] = m + n
-
-        runner = ConjectureRunner(test, settings=TEST_SETTINGS)
-        runner.cached_test_function([200, 0])
-        runner.cached_test_function([0, 200])
-
-        try:
-            runner.optimise_targets()
-        except RunIsComplete:
-            pass
-
-        assert runner.best_observed_targets["m"] == 255
-        assert runner.best_observed_targets["n"] == 255
-        assert runner.best_observed_targets["m + n"] == 256
-
-
 def test_does_not_shrink_multiple_bugs_when_told_not_to():
     def test(data):
         m = data.draw_bits(8)
